rename packages to highlight their testing purpose
