+++ internal/testcommon/k8stest/telemetrygen.go
@@ -1,7 +1,7 @@
 // Copyright The OpenTelemetry Authors
 // SPDX-License-Identifier: Apache-2.0
 
-package k8s // import "github.com/open-telemetry/opentelemetry-collector-contrib/internal/k8stest"
+package k8stest // import "github.com/open-telemetry/opentelemetry-collector-contrib/internal/k8stest"
 
 import (
 	"bytes"
